requirements.txt
openai==1.50.2
qq_botpy==1.2.1
